ivy.el (ivy-occur-flush): Add and bind to "F"
Use case 1:

Suppose we want to grep for "defun.*occur", but leave out
private ("--") functions.

We can do it with:

    "defun occur ! --" C-c C-o

But now we can also do it with:

    "defun occur" C-c C-o F "--" RET

The latter approach is more flexible.

Use case 2:

Suppose we're looking for a function that starts with "vi". So use
`counsel-describe-function' "vi" C-c C-o.

- We get 89 cands.
- F "view" RET to get 30 cands.
- F "viet" RET to get 25 cands.
- F "vim" RET to get 11 cands.
- now it's easy to go through the docs of the functions that are left.